requirements.txt
aiofiles
annotated-doc
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
bcrypt
cffi
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
h11
idna
motor
passlib
pyasn1
pycparser
pydantic
pydantic_core
pymongo
python-jose
python-multipart
rsa
six
starlette
typing-inspection
typing_extensions
uvicorn
python-dotenv
fastapi-mail
python-multipart